refactor(xras): shared vocabulary out of whoever needed it first
## Summary

Two new sibling modules under `handlers/`, and **no handler module imports another
handler module** any more. `update.py` was importing fifteen symbols from three
siblings — the import graph encoded build order, not domain structure. Nothing
about the domain says Update depends on New.

- **`_fields.py`** — one wire field in, one value out, reporting into `errs`:
  `title`, `abstract`, `parse_action_begin_date`, `parse_action_end_date`,
  `transaction_amount`, `resource_comment`, `resolve_resource`, `plan_contracts`.
- **`_allocations.py`** — finding, dating and marking the allocations an action
  acts on: `account_is_active`, `effective_end_date`, `latest_allocation`,
  `account_for_resource`, `new_allocation_end_date`, `clamp_start_to_commission`,
  `auth_at_panel_meeting`, `mark_panel_authorised`, and the two create policies.

### The create policies now have names

`create_window_from_project_history` (Supplement, Adjustment — today plus the
project's own history) and `create_window_from_action_dates` (New, Update — the
action's dates, start clamped to commissioning). Two call sites each.

This is the specific thing that prevents a recurrence of the bug fixed two commits
ago: two call sites of one function cannot drift, where two copies of thirty lines
can and did.

Both policies are legacy's and **both must survive**. Naming them is not a step
toward merging them.

### Two ⚠️ that used to be enforced by living in different files

`_allocations.py` opens by naming them, because the module now holds both halves:

- `account_is_active` filters hard, `account_for_resource` is deliberately
  unfiltered — so a Supplement lands on a decommissioned resource where an
  Extension would skip it.
- `auth_at_panel_mtg` reaches a row two different ways: marked after the fact on a
  CREATE row, passed as a parameter on a SUPPLEMENT row, and not at all on an
  ADJUSTMENT row.

`auth_at_panel_meeting` also gains a warning it did not have: it must be called
during assembly. Its second arm reads `project.allocation_type`, and Update
*writes* that column through `project.update()`, which flushes. Evaluating it
lazily from the execute phase would read back the type the action just installed.
No test would catch that.

`plan_contracts` gets a home, retiring `update.py`'s function-body
`from .new import _plan_contracts` — executed on every call.

### Test Results

⚠️ **845 xras tests pass with ZERO test edits.** The moved names are re-exported
from their old modules (`# noqa: F401`) precisely so the suite has to pass
untouched — which makes it evidence that the move was pure rather than a net that
was adjusted to fit. The re-exports come out in the next commit, with the test
imports repointed.

Leak check → 0.
